requirements.txt
PyYAML>=6.0.1
pydantic[email]>=2.7
httpx>=0.27
APScheduler>=3.10
Jinja2>=3.1
loguru>=0.7

# Dev dependencies
pytest>=8.0
pytest-cov>=5.0
ruff>=0.6
mypy>=1.10
types-PyYAML>=6.0.12.20240808
